Split out face vertex/index generation and move everything into folder
