+++ src/accessors/displacements/indices.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+namespace BspParser::Accessors::Internal {}
