+++ src/accessors/displacements/vertices.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+namespace BspParser::Accessors::Internal {}
