+++ src/accessors/face-accessors.hpp
@@ -1,18 +1,12 @@
 #pragma once
 
-#include "bsp.hpp"
-#include "structs/geometry.hpp"
-#include "structs/models.hpp"
+#include "../bsp.hpp"
+#include "../structs/geometry.hpp"
+#include "../structs/models.hpp"
+#include "./vertex.hpp"
 #include <functional>
 
 namespace BspParser::Accessors {
-  struct Vertex {
-    Structs::Vector position;
-    Structs::Vector normal;
-    Structs::Vector4 tangent;
-    Structs::Vector2 uv;
-  };
-
   void iterateModels(
     const Bsp& bsp,
     const std::function<void(const Structs::Model& model, const std::vector<Bsp::PhysModel>& physicsModels)>& iteratee
@@ -32,23 +26,25 @@ namespace BspParser::Accessors {
   /**
    * Returns the number of vertices that will be generated by generateFaceVertices.
    * Useful for precomputing the size of vertex buffer needed to reduce allocations.
+   * @param bsp
    * @param face
    * @param surfaceEdges
    * @return Number of vertices for the face (0 if not drawable or a displacement).
    */
-  size_t getFaceVertexCount(const Structs::Face& face, std::span<const int32_t> surfaceEdges);
+  size_t getVertexCount(const Bsp& bsp, const Structs::Face& face, std::span<const int32_t> surfaceEdges);
 
   /**
    * Returns the number of indices that will be generated by generateFaceTriangleList.
    * Useful for precomputing the size of index buffer needed to reduce allocations.
+   * @param bsp
    * @param face
    * @param surfaceEdges
    * @return Number of indices for the face (0 if not drawable or a displacement).
    */
-  size_t getFaceTriangleListIndexCount(const Structs::Face& face, std::span<const int32_t> surfaceEdges);
+  size_t getTriangleListIndexCount(const Bsp& bsp, const Structs::Face& face, std::span<const int32_t> surfaceEdges);
 
   /**
-   * Calls iteratee once for each unique vertex in the face's edges.
+   * Calls iteratee once for each unique vertex in the face's edges or displacement.
    * Generates normals, tangents and UVs.
    * @param bsp
    * @param face Face to generate vertices for.
@@ -57,7 +53,7 @@ namespace BspParser::Accessors {
    * @param surfaceEdges Surface edges referenced by the face.
    * @param iteratee Function to call with each generated vertex.
    */
-  void generateFaceVertices(
+  void generateVertices(
     const Bsp& bsp,
     const Structs::Face& face,
     const Structs::Plane& plane,
@@ -69,11 +65,13 @@ namespace BspParser::Accessors {
   /**
    * Calls iteratee once for each triangle forming a mesh which triangulates the given face.
    * Indices start from 0 and index into the vertices generated by generateFaceVertices.
+   * @param bsp
    * @param face Face to generate a triangle list for.
    * @param surfaceEdges Surface edges of the face to generate a triangle list for.
    * @param iteratee Called with each triplet of indices defining a triangle with clockwise winding.
    */
-  void generateFaceTriangleList(
+  void generateTriangleListIndices(
+    const Bsp& bsp,
     const Structs::Face& face,
     std::span<const int32_t> surfaceEdges,
     const std::function<void(int32_t i0, int32_t i1, int32_t i2)>& iteratee
