+++ src/accessors/texture-accessors.cpp
