+++ src/accessors/vertex.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "../structs/common.hpp"
+
+namespace BspParser::Accessors {
+  struct Vertex {
+    Structs::Vector position;
+    Structs::Vector normal;
+    Structs::Vector4 tangent;
+    Structs::Vector2 uv;
+  };
+}
