--- src/displacement-accessors.cpp
@@ -1,1 +1,0 @@
-#include "displacement-accessors.hpp"
