--- src/displacement-accessors.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace BspParser::Accessors {
-  // TODO: Design displacement accessor API
-  // need to support triangulation to tri lists, and possibly height fields for PhysX but that might be more effort than it's worth
-}
